add dimension getters
